// psa.service.sampletrackingservice/src/filters/model/segmentElement.ts
/*
 * SPDX-FileCopyrightText: 2021 Helmholtz-Zentrum für Infektionsforschung GmbH (HZI) <PiaPost@helmholtz-hzi.de>
 *
 * SPDX-License-Identifier: AGPL-3.0-or-later
 */

export interface SegmentElement {
  name: string;
  attributes: { key: string; value: string }[];
  children: SegmentElement[];
}